fix(codecov): Remove integrated picker and simplify selector component
